Added READMEs for each demo
The demos here are stripped down to make them easy to show on the screen,
but this makes it hard to remember how they work -- and makes it hard for
someone just following along later to understand the demo.

THe READMEs outline how to run the demos and what they're supposed to demo
as both a reminder to me and help for anyone else.